Split handlers.ts phase 2: extract run + plan-run handlers (warren-a2b4 / pl-9088 step 2)
Extract run and plan-run HTTP handlers from src/server/handlers/index.ts
into src/server/handlers/runs.ts and src/server/handlers/plan-runs.ts —
the second pass over the handlers monolith (phase 1 split out plot +
plot-plan-runs; see mx-3df5c5 / warren-48de for the playbook).

Production changes
- src/server/handlers/runs.ts (789 lines, new): listRunsHandler,
  getRunHandler, createRunHandler, postRunMessageHandler,
  steerRunHandler, cancelRunHandler, previewLoginHandler,
  previewTeardownHandler, streamRunEventsHandler,
  listCostAnalyticsHandler, plus the streaming plumbing (bridgeAbort,
  asNdjsonStream, eventToNdjson) consumed by both run and plan-run
  NDJSON tails.
- src/server/handlers/plan-runs.ts (484 lines, new): createPlanRunHandler,
  listPlanRunsHandler, getPlanRunHandler, cancelPlanRunHandler,
  streamPlanRunEventsHandler + tailPlanRunEvents.
- src/server/handlers/index.ts (2391 → 1194 lines): keeps ROUTE_TABLE,
  shared parsing helpers (readJsonBody[OrEmpty], requireString,
  optionalString, requireParam, parseBoolean, parseNonNegativeInt,
  assertPlotIdDispatchable, defaultSpawn — all exported), agent /
  project / burrow / worker / meta handlers, and the auth gate.

ROUTE_TABLE stays in index.ts as the canonical wire surface so
gen:docs / gen:openapi keep parsing one file (mx-3df5c5 step 1).

Test changes
- src/server/handlers/runs.test.ts (988 lines, new): extracts the
  POST /runs (spawn flow + plot_id validation + interactive mode),
  POST /runs/:id/messages, and GET /runs/:id/events describes from
  src/server/handlers.test.ts. Shared helpers are duplicated since
  the residual handlers.test.ts still needs them.
- src/server/handlers/plan-runs.test.ts (1244 → 1318 lines, moved):
  relocated from src/server/handlers.plan-runs.test.ts. The
  noExcessiveLinesPerFunction override is removed; the monolithic
  POST /plan-runs describe is split into three siblings (core +
  'Plot integration (warren-b89f)' + 'plot_id validation (warren-bae5)')
  with a shared setupPlanRunFixture() helper so each sibling fits
  under the 500-line per-function cap.
- src/server/handlers.test.ts (2455 → 1627 lines): residual describes
  for refresh / warren-config / agents / triggers / project agents
  refresh / projectId filter / seeds / brainstorm.

Ratchet updates
- biome.json: drop handlers.plan-runs.test.ts from the
  noExcessiveLinesPerFunction override list; add handlers/runs.ts +
  handlers/plan-runs.ts to the cognitive-complexity override block.
- scripts/file-size-budgets.json: lower handlers/index.ts (2391 →
  1194) + handlers.test.ts (2455 → 1627); drop the old
  handlers.plan-runs.test.ts entry; add budgets for the new files.
- AGENTS.md: update the dotted-grouping example to the new path.

Quality gates green (bun run check:all).
