+++ biome.json
@@ -67,8 +67,7 @@
 				"src/runs/reap.test.ts",
 				"src/runs/reap.ts",
 				"src/runs/spawn.test.ts",
-				"src/runs/stream.test.ts",
-				"src/server/handlers.plan-runs.test.ts"
+				"src/runs/stream.test.ts"
 			],
 			"linter": {
 				"rules": {
@@ -125,8 +124,10 @@
 				"src/server/bridges.ts",
 				"src/server/errors.ts",
 				"src/server/handlers/index.ts",
+				"src/server/handlers/plan-runs.ts",
 				"src/server/handlers/plot-plan-runs.ts",
 				"src/server/handlers/plots.ts",
+				"src/server/handlers/runs.ts",
 				"src/server/integration.multi-worker.test.ts",
 				"src/server/main.ts",
 				"src/server/server.ts",
