+++ scripts/file-size-budgets.json
@@ -40,13 +40,15 @@
 		"src/runs/stream.ts": 911,
 		"src/server/bridges.test.ts": 647,
 		"src/server/bridges.ts": 655,
-		"src/server/handlers.plan-runs.test.ts": 1244,
 		"src/server/handlers.preview.test.ts": 857,
-		"src/server/handlers.test.ts": 2455,
-		"src/server/handlers/index.ts": 2391,
+		"src/server/handlers.test.ts": 1627,
+		"src/server/handlers/index.ts": 1194,
+		"src/server/handlers/plan-runs.test.ts": 1318,
 		"src/server/handlers/plot-plan-runs.test.ts": 640,
 		"src/server/handlers/plots.test.ts": 3726,
 		"src/server/handlers/plots.ts": 1652,
+		"src/server/handlers/runs.test.ts": 988,
+		"src/server/handlers/runs.ts": 789,
 		"src/server/integration.multi-worker.test.ts": 553,
 		"src/server/main.ts": 700,
 		"src/server/server.test.ts": 536,
