+++ src/server/handlers/index.ts
@@ -26,14 +26,7 @@
  * progress channel onto this route.
  */
 
-import { join } from "node:path";
-import type {
-	Burrow,
-	BurrowKind,
-	BurrowState,
-	HttpBurrowListFilter,
-	MessagePriority,
-} from "@os-eco/burrow-cli";
+import type { Burrow, BurrowKind, BurrowState, HttpBurrowListFilter } from "@os-eco/burrow-cli";
 import { withTransportMapping } from "../../burrow-client/client.ts";
 import { fanOutAcrossWorkers } from "../../burrow-client/fanout.ts";
 import { NotFoundError, ValidationError } from "../../core/errors.ts";
@@ -53,45 +46,21 @@ import {
 	type DiagnosticCheck,
 } from "../../diagnostics/checks.ts";
 import { VERSION } from "../../index.ts";
-import {
-	PlanHasNoOpenChildrenError,
-	ProjectLacksPlotError,
-	ProjectLacksSeedsError,
-} from "../../plan-runs/errors.ts";
-import {
-	defaultPlanRunPlotAppender,
-	emitPlanRunDispatchedToPlot,
-} from "../../plan-runs/plot-appender.ts";
+import { ProjectLacksSeedsError } from "../../plan-runs/errors.ts";
 import { isValidPlotIdFormat, PlotIdInvalidError, PlotIdNotFoundError } from "../../plots/index.ts";
 import { createRunPreviewsRepo, DEFAULT_MAX_LIVE } from "../../preview/eviction.ts";
 import { DEFAULT_PREVIEW_PORT_RANGE, PreviewPortAllocator } from "../../preview/port-allocator.ts";
-import { teardownPreview } from "../../preview/teardown.ts";
 import type { SpawnFn, SpawnOptions, SpawnResult } from "../../projects/clone.ts";
 import { addProject, deleteProject, listProjects, refreshProject } from "../../projects/index.ts";
 import { type AgentSource, readAgentSource } from "../../registry/builtins/index.ts";
 import { CanopyClient } from "../../registry/canopy.ts";
-import { readProviderFrontmatter } from "../../registry/index.ts";
 import {
 	type RefreshProjectResult,
 	refreshAgentRegistry,
 	refreshProjectAgents,
 } from "../../registry/refresh.ts";
-import {
-	appendUserMessage,
-	buildCostAnalytics,
-	buildInteractivePrompt,
-	type CostAnalyticsRow,
-	cancelRun,
-	defaultPlotContextReader,
-	hydrateRunsUsage,
-	hydrateRunUsage,
-	resolveDispatcherHandle,
-	spawnInteractiveTurn,
-	spawnRun,
-	steerRun,
-	tailRunEvents,
-} from "../../runs/index.ts";
-import { showPlan, showSeed } from "../../seeds-cli/index.ts";
+import { spawnRun } from "../../runs/index.ts";
+import { showSeed } from "../../seeds-cli/index.ts";
 import { buildTriggerSummaries, parseCron, resolveCronPrompt } from "../../triggers/index.ts";
 import {
 	type CronTrigger,
@@ -99,8 +68,15 @@ import {
 	type LoadedWarrenConfig,
 	loadWarrenConfig,
 } from "../../warren-config/index.ts";
-import { jsonResponse, ndjsonResponse } from "../response.ts";
+import { jsonResponse } from "../response.ts";
 import type { Route, RouteContext, RouteHandler, ServerDeps } from "../types.ts";
+import {
+	cancelPlanRunHandler,
+	createPlanRunHandler,
+	getPlanRunHandler,
+	listPlanRunsHandler,
+	streamPlanRunEventsHandler,
+} from "./plan-runs.ts";
 import { createPlotPlanRunHandler } from "./plot-plan-runs.ts";
 import {
 	answerPlotQuestionHandler,
@@ -119,6 +95,18 @@ import {
 	renamePlotHandler,
 	syncPlotHandler,
 } from "./plots.ts";
+import {
+	cancelRunHandler,
+	createRunHandler,
+	getRunHandler,
+	listCostAnalyticsHandler,
+	listRunsHandler,
+	postRunMessageHandler,
+	previewLoginHandler,
+	previewTeardownHandler,
+	steerRunHandler,
+	streamRunEventsHandler,
+} from "./runs.ts";
 
 /**
  * Run a command via `Bun.spawn` and return the SpawnResult shape the
@@ -205,14 +193,14 @@ export function requireParam(ctx: RouteContext, key: string): string {
 	return value;
 }
 
-function parseBoolean(raw: string | null, label: string): boolean | undefined {
+export function parseBoolean(raw: string | null, label: string): boolean | undefined {
 	if (raw === null) return undefined;
 	if (raw === "true" || raw === "1") return true;
 	if (raw === "false" || raw === "0") return false;
 	throw new ValidationError(`${label} must be 'true'/'1' or 'false'/'0'; got '${raw}'`);
 }
 
-function parseNonNegativeInt(raw: string | null, label: string): number | undefined {
+export function parseNonNegativeInt(raw: string | null, label: string): number | undefined {
 	if (raw === null) return undefined;
 	const n = Number.parseInt(raw, 10);
 	if (!Number.isFinite(n) || n < 0 || String(n) !== raw) {
@@ -254,7 +242,7 @@ function withAgentSource(row: AgentRow): AgentRow & { source: AgentSource } {
  *
  * warren-bae5 / pl-5310 step 2 — fold-in of warren-a353.
  */
-async function assertPlotIdDispatchable(input: {
+export async function assertPlotIdDispatchable(input: {
 	readonly plotId: string | undefined;
 	readonly plotResolver?: import("../../plots/index.ts").PlotResolver;
 }): Promise<void> {
@@ -868,1196 +856,6 @@ function parseDrainFlag(body: Record<string, unknown>): boolean {
 		throw new ValidationError("field 'drain' must be a boolean");
 	}
 	return raw;
-}
-
-/* ----------------------------------------------------------------------- */
-/* Runs (§8.1)                                                              */
-/* ----------------------------------------------------------------------- */
-
-function parseRunsSort(ctx: { url: URL }): { sort: "started" | "cost"; dir: "asc" | "desc" } {
-	const rawSort = ctx.url.searchParams.get("sort");
-	const rawDir = ctx.url.searchParams.get("dir");
-	let sort: "started" | "cost" = "started";
-	if (rawSort !== null) {
-		if (rawSort !== "started" && rawSort !== "cost") {
-			throw new ValidationError("?sort must be 'started' or 'cost'");
-		}
-		sort = rawSort;
-	}
-	let dir: "asc" | "desc" = "desc";
-	if (rawDir !== null) {
-		if (rawDir !== "asc" && rawDir !== "desc") {
-			throw new ValidationError("?dir must be 'asc' or 'desc'");
-		}
-		dir = rawDir;
-	}
-	return { sort, dir };
-}
-
-/**
- * Parse `?limit` / `?offset` for the runs list (warren-ee50 / pl-b0c0
- * step 1). Defaults preserve the historical 100-row window so existing
- * consumers stay byte-compatible; `limit` is clamped to a sane range
- * to keep a paginated UI honest and the server cheap. `offset` is
- * non-negative; the caller is responsible for stitching pages on its
- * end (no opaque cursor — the predicates here are stable + the
- * `runs.id` tiebreaker in `orderByClause` keeps the order total).
- */
-function parseRunsPagination(ctx: { url: URL }): { limit: number; offset: number } {
-	const rawLimit = ctx.url.searchParams.get("limit");
-	const rawOffset = ctx.url.searchParams.get("offset");
-	let limit = 100;
-	if (rawLimit !== null) {
-		const n = Number.parseInt(rawLimit, 10);
-		if (!Number.isFinite(n) || n <= 0 || String(n) !== rawLimit) {
-			throw new ValidationError(`?limit must be a positive integer; got '${rawLimit}'`);
-		}
-		if (n > 500) throw new ValidationError("?limit must be ≤ 500");
-		limit = n;
-	}
-	let offset = 0;
-	if (rawOffset !== null) {
-		const n = Number.parseInt(rawOffset, 10);
-		if (!Number.isFinite(n) || n < 0 || String(n) !== rawOffset) {
-			throw new ValidationError(`?offset must be a non-negative integer; got '${rawOffset}'`);
-		}
-		offset = n;
-	}
-	return { limit, offset };
-}
-
-function listRunsHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const project = ctx.url.searchParams.get("project");
-		const agent = ctx.url.searchParams.get("agent");
-		if (project !== null && agent !== null) {
-			throw new ValidationError("filter by either ?project=... or ?agent=..., not both");
-		}
-		const order = parseRunsSort(ctx);
-		const page = parseRunsPagination(ctx);
-		const listOpts = { ...order, ...page };
-		const rows =
-			project !== null
-				? await deps.repos.runs.listByProject(project, listOpts)
-				: agent !== null
-					? await deps.repos.runs.listByAgent(agent, listOpts)
-					: await deps.repos.runs.listAll(listOpts);
-		// warren-ab18: surface in-events cost for terminal runs whose
-		// bridge died before the final checkpoint landed.
-		const runs = await hydrateRunsUsage(rows, deps.repos.events);
-		// warren-ee50 / pl-b0c0 step 1: aggregate the full filtered set so
-		// the Runs page can show all-time totals next to a paginated table.
-		const aggFilter = {
-			...(project !== null ? { projectId: project } : {}),
-			...(agent !== null ? { agentName: agent } : {}),
-		};
-		const agg = await deps.repos.runs.aggregate(aggFilter);
-		return jsonResponse(200, {
-			runs,
-			total: agg.total,
-			costTotalUsd: agg.costTotalUsd,
-			costPricedCount: agg.costPricedCount,
-			limit: page.limit,
-			offset: page.offset,
-		});
-	};
-}
-
-function getRunHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const id = requireParam(ctx, "id");
-		const row = await deps.repos.runs.require(id);
-		// warren-ab18: same compute-on-read fallback as the list handler
-		// so the RunDetail page shows cost for ghost / reboot-orphaned runs.
-		const run = await hydrateRunUsage(row, deps.repos.events);
-		return jsonResponse(200, run);
-	};
-}
-
-/**
- * `GET /analytics/cost?from=&to=&projectId=` (warren-cf63 / pl-b0c0 step 6).
- *
- * Window defaults to the last 30 days when neither bound is supplied so
- * a fresh install renders a useful chart without operator setup. Both
- * bounds and `projectId` are validated lightly — a malformed date is a
- * 400 because the lexicographic ISO8601 compare in `listForAnalytics`
- * would silently produce surprising results otherwise.
- */
-function listCostAnalyticsHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const projectId = ctx.url.searchParams.get("projectId") ?? undefined;
-		const from = parseAnalyticsDateBound(ctx, "from");
-		const to = parseAnalyticsDateBound(ctx, "to");
-		const defaultFrom = (() => {
-			if (from !== undefined) return from;
-			if (to !== undefined) return undefined;
-			const d = new Date(Date.now() - 30 * 24 * 60 * 60 * 1000);
-			return d.toISOString();
-		})();
-		const filter: { projectId?: string; from?: string; to?: string } = {};
-		if (projectId !== undefined) filter.projectId = projectId;
-		if (defaultFrom !== undefined) filter.from = defaultFrom;
-		if (to !== undefined) filter.to = to;
-		const rowsRaw = await deps.repos.runs.listForAnalytics(filter);
-		// Hydrate so terminal runs with bridge-died cost still count.
-		const rows = await hydrateRunsUsage(rowsRaw, deps.repos.events);
-		const planByRun = new Map<string, string>();
-		if (rows.length > 0) {
-			const joined = await deps.repos.planRuns.resolvePlanForRunIds(rows.map((r) => r.id));
-			for (const j of joined) planByRun.set(j.runId, j.planId);
-		}
-		const analyticsRows: CostAnalyticsRow[] = rows.map((r) => {
-			const { provider, model } = extractProviderModel(r.renderedAgentJson);
-			return {
-				runId: r.id,
-				projectId: r.projectId,
-				agentName: r.agentName,
-				plotId: r.plotId,
-				planId: planByRun.get(r.id) ?? null,
-				planRunId: null,
-				provider: provider ?? null,
-				model: model ?? null,
-				costUsd: r.costUsd,
-				startedAt: r.startedAt,
-			};
-		});
-		const analytics = buildCostAnalytics(analyticsRows);
-		return jsonResponse(200, {
-			filter: {
-				projectId: projectId ?? null,
-				from: defaultFrom ?? null,
-				to: to ?? null,
-			},
-			...analytics,
-		});
-	};
-}
-
-function parseAnalyticsDateBound(ctx: { url: URL }, name: "from" | "to"): string | undefined {
-	const raw = ctx.url.searchParams.get(name);
-	if (raw === null || raw === "") return undefined;
-	const d = new Date(raw);
-	if (Number.isNaN(d.getTime())) {
-		throw new ValidationError(`?${name} must be an ISO8601 date or datetime`);
-	}
-	return d.toISOString();
-}
-
-function extractProviderModel(rendered: unknown): { provider?: string; model?: string } {
-	if (rendered === null || typeof rendered !== "object") return {};
-	const fm = (rendered as { frontmatter?: unknown }).frontmatter;
-	if (fm === null || fm === undefined || typeof fm !== "object") return {};
-	return readProviderFrontmatter(fm as Readonly<Record<string, unknown>>);
-}
-
-/**
- * Parse the optional `mode` body field on `POST /runs` (pl-0344 step 4 /
- * warren-b3b9). Defaults to `'batch'` so existing single-shot dispatch
- * callers are byte-identical. `'interactive'` opts in to the
- * respawn-per-turn primitive (`src/runs/interactive.ts`); when set,
- * `plotId` is required and `interactiveAgent` may override `agent`.
- */
-function parseRunMode(body: Record<string, unknown>): "batch" | "interactive" {
-	const raw = body.mode;
-	if (raw === undefined || raw === null) return "batch";
-	if (raw !== "batch" && raw !== "interactive") {
-		throw new ValidationError(
-			`field 'mode' must be 'batch' or 'interactive'; got ${JSON.stringify(raw)}`,
-		);
-	}
-	return raw;
-}
-
-function createRunHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const body = await readJsonBody(ctx);
-		const mode = parseRunMode(body);
-		const ref = optionalString(body, "ref");
-		const providerOverride = optionalString(body, "providerOverride");
-		const modelOverride = optionalString(body, "modelOverride");
-		const seedId = optionalString(body, "seedId");
-		const plotId = optionalString(body, "plotId");
-		const dispatcherHandle = optionalString(body, "dispatcherHandle");
-		// `interactiveAgent` overrides `agent` when mode='interactive' (pl-0344
-		// step 4 / warren-b3b9). The field is dedicated so a UI surface that
-		// always sends both `agent` (batch default) and `interactiveAgent`
-		// (brainstorm/planner preset) can flip mode without re-keying.
-		const interactiveAgent = optionalString(body, "interactiveAgent");
-		const agentName =
-			mode === "interactive" && interactiveAgent !== undefined
-				? interactiveAgent
-				: requireString(body, "agent");
-		const projectId = requireString(body, "project");
-		const prompt = requireString(body, "prompt");
-
-		// warren-bae5 / pl-5310 step 2: validate plot_id BEFORE handing off
-		// to spawnRun so a malformed or non-existent plot_id fails at the
-		// operator-facing edge instead of silently no-opping at the
-		// host-side defaultPlotAppender. Empty string is normalized to "no
-		// plot bound" (matches spawnRun's posture and the createPlanRun gate).
-		await assertPlotIdDispatchable({ plotId, plotResolver: deps.plotResolver });
-
-		// Interactive runs REQUIRE a plot_id — the primitive is meaningless
-		// without a Plot to bind the conversation to. Validated AFTER format
-		// + existence so the more-specific format error still wins when both
-		// apply.
-		if (mode === "interactive" && (plotId === undefined || plotId === "")) {
-			throw new ValidationError(
-				"plotId is required when mode='interactive'; interactive runs bind to a Plot",
-				{
-					recoveryHint: "either pass plotId on POST /runs, or omit `mode` to dispatch a batch run",
-				},
-			);
-		}
-
-		// Compose the dispatch prompt. Batch runs ship the operator-supplied
-		// prompt verbatim. Interactive first-turn dispatches wrap it in the
-		// same <plot_context> + <user_message> envelope the follow-up turn
-		// builder (`spawnInteractiveTurn`) emits, so the agent sees a uniform
-		// prompt shape across turn 1 and turn N. Context load is best-effort
-		// — a torn `.plot/.index.db` or vanished events file falls back to
-		// the user message alone.
-		let dispatchedPrompt = prompt;
-		if (mode === "interactive" && plotId !== undefined && plotId !== "") {
-			const project = await deps.repos.projects.require(projectId);
-			const handle = resolveDispatcherHandle(dispatcherHandle);
-			let context = null;
-			try {
-				context = await defaultPlotContextReader.read({
-					plotDir: join(project.localPath, ".plot"),
-					plotId,
-					historyTail: 0,
-					handle,
-				});
-			} catch {
-				// Best-effort — first-turn dispatches still proceed without
-				// the envelope. spawnInteractiveTurn captures the same
-				// degradation as a system event; on first turn we don't yet
-				// have a run row to attach the warning to.
-			}
-			dispatchedPrompt = buildInteractivePrompt(context, prompt);
-		}
-
-		const result = await spawnRun({
-			repos: deps.repos,
-			burrowClientPool: deps.burrowClientPool,
-			agentName,
-			projectId,
-			prompt: dispatchedPrompt,
-			mode,
-			...(body.metadata !== undefined ? { metadata: body.metadata } : {}),
-			...(deps.now !== undefined ? { now: deps.now } : {}),
-			projectsConfig: deps.projectsConfig,
-			projectSpawn: deps.spawn ?? defaultSpawn,
-			...(ref !== undefined ? { ref } : {}),
-			...(providerOverride !== undefined ? { providerOverride } : {}),
-			...(modelOverride !== undefined ? { modelOverride } : {}),
-			...(seedId !== undefined ? { seedId } : {}),
-			...(plotId !== undefined ? { plotId } : {}),
-			...(dispatcherHandle !== undefined ? { dispatcherHandle } : {}),
-			...(deps.warrenConfigs !== undefined ? { warrenConfigs: deps.warrenConfigs } : {}),
-			...(deps.runBranchPrefixDefault !== undefined
-				? { runBranchPrefixDefault: deps.runBranchPrefixDefault }
-				: {}),
-			...(deps.seedsCli !== undefined ? { seedsCli: deps.seedsCli } : {}),
-		});
-
-		// First-turn interactive dispatch: append the raw user_message event
-		// onto the new run so the events stream reflects the conversation
-		// from turn 0. `spawnInteractiveTurn` (follow-up turns) does the same.
-		if (mode === "interactive") {
-			await appendUserMessage({
-				repos: deps.repos,
-				runId: result.run.id,
-				message: prompt,
-				handle: resolveDispatcherHandle(dispatcherHandle),
-				...(deps.now !== undefined ? { now: deps.now() } : {}),
-			});
-		}
-
-		// Hand off to the bridge so events start flowing into warren.events
-		// — without this the dispatched run would emit events into burrow
-		// but the warren wire would never see them.
-		deps.bridges.start(result.run.id, result.burrowRun.id, result.burrow.id);
-		return jsonResponse(201, {
-			run: result.run,
-			burrow: { id: result.burrow.id, workspacePath: result.burrow.workspacePath },
-		});
-	};
-}
-
-/**
- * `POST /runs/:id/messages` — send a follow-up user turn on an
- * interactive conversation (pl-0344 step 4 / warren-b3b9).
- *
- * `:id` is the conversation handle — any prior interactive run row that
- * shares the same plotId works (the handler resolves the plot context
- * from disk, not from this row). Returns **202 Accepted** with the
- * freshly-spawned turn row + burrow descriptor + user_message event:
- * the dispatch is async (the agent reply lands later as an
- * `agent_message` event captured at reap), and 202 is the canonical
- * "queued for processing, see events stream for completion" shape.
- *
- * Body: `{message: string, dispatcherHandle?: string, providerOverride?,
- * modelOverride?, ref?}`. Field shape mirrors `POST /runs` so a UI
- * surface can reuse its dispatch plumbing.
- *
- * Errors:
- *   - 400 `validation_error` — missing/empty message, prior run is not
- *     mode='interactive', prior run has no plot_id, message is empty.
- *   - 404 `not_found` — prior run id doesn't exist.
- *
- * The follow-up turn is registered with the bridge the same way `POST
- * /runs` does, so its events flow onto warren's stream immediately.
- */
-function postRunMessageHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const id = requireParam(ctx, "id");
-		const body = await readJsonBody(ctx);
-		const message = requireString(body, "message");
-		const dispatcherHandle = optionalString(body, "dispatcherHandle");
-		const providerOverride = optionalString(body, "providerOverride");
-		const modelOverride = optionalString(body, "modelOverride");
-		const ref = optionalString(body, "ref");
-
-		const result = await spawnInteractiveTurn({
-			runId: id,
-			message,
-			repos: deps.repos,
-			burrowClientPool: deps.burrowClientPool,
-			trigger: "interactive",
-			projectsConfig: deps.projectsConfig,
-			projectSpawn: deps.spawn ?? defaultSpawn,
-			...(ref !== undefined ? { ref } : {}),
-			...(providerOverride !== undefined ? { providerOverride } : {}),
-			...(modelOverride !== undefined ? { modelOverride } : {}),
-			...(dispatcherHandle !== undefined ? { dispatcherHandle } : {}),
-			...(deps.now !== undefined ? { now: deps.now } : {}),
-			...(deps.warrenConfigs !== undefined ? { warrenConfigs: deps.warrenConfigs } : {}),
-			...(deps.runBranchPrefixDefault !== undefined
-				? { runBranchPrefixDefault: deps.runBranchPrefixDefault }
-				: {}),
-			...(deps.seedsCli !== undefined ? { seedsCli: deps.seedsCli } : {}),
-		});
-
-		deps.bridges.start(result.turn.run.id, result.turn.burrowRun.id, result.turn.burrow.id);
-
-		return jsonResponse(202, {
-			run: result.turn.run,
-			burrow: {
-				id: result.turn.burrow.id,
-				workspacePath: result.turn.burrow.workspacePath,
-			},
-			userMessageEvent: {
-				id: result.userMessageEvent.id,
-				runId: result.userMessageEvent.runId,
-				seq: result.userMessageEvent.burrowEventSeq,
-				ts: result.userMessageEvent.ts,
-				kind: result.userMessageEvent.kind,
-			},
-			priorRunId: result.priorRun.id,
-			plotContextDegraded: result.plotContextDegraded,
-		});
-	};
-}
-
-function steerRunHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const id = requireParam(ctx, "id");
-		const body = await readJsonBody(ctx);
-		const result = await steerRun({
-			runId: id,
-			body: requireString(body, "body"),
-			repos: deps.repos,
-			burrowClientPool: deps.burrowClientPool,
-			broker: deps.broker,
-			...(optionalString(body, "priority") !== undefined
-				? { priority: optionalString(body, "priority") as MessagePriority }
-				: {}),
-			...(optionalString(body, "fromActor") !== undefined
-				? { fromActor: optionalString(body, "fromActor") as string }
-				: {}),
-			...(deps.now !== undefined ? { now: deps.now } : {}),
-		});
-		return jsonResponse(200, { message: result.message });
-	};
-}
-
-function cancelRunHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const id = requireParam(ctx, "id");
-		const body = await readJsonBodyOrEmpty(ctx);
-		const reason = body !== null ? optionalString(body, "reason") : undefined;
-		const result = await cancelRun({
-			runId: id,
-			repos: deps.repos,
-			burrowClientPool: deps.burrowClientPool,
-			broker: deps.broker,
-			...(reason !== undefined ? { reason } : {}),
-			...(deps.now !== undefined ? { now: deps.now } : {}),
-			...(deps.autoOpenPr !== undefined ? { autoOpenPr: deps.autoOpenPr } : {}),
-		});
-		return jsonResponse(200, {
-			state: result.state,
-			alreadyTerminal: result.alreadyTerminal,
-			burrowRun: result.burrowRun,
-		});
-	};
-}
-
-/**
- * `GET /runs/:id/preview/login?token=<bearer>&redirect=<absolute-url>`
- * (R-19 / SPEC §11.L, warren-8a10; path-mode redirect warren-edff;
- * per-run cookie name warren-63e1).
- *
- * The signed-cookie handshake the preview proxy depends on. A browser
- * hitting a preview origin directly can't carry an Authorization header,
- * so the operator opens this URL on the warren host, the handler
- * validates the bearer in the query, sets a scoped `warren_preview*`
- * cookie, and 302s to the preview.
- *
- *   - **Subdomain mode** (`deps.previewMode === "subdomain"`): cookie name
- *     `warren_preview`, `Domain=.<host>; Path=/`; redirect must be
- *     `https://run-<id>.<previewHost>/...`.
- *   - **Path mode** (default; `deps.previewMode === "path"`): cookie name
- *     `warren_preview_<runId>` (per-run literal suffix, warren-63e1),
- *     `Path=/` with no `Domain`; redirect must be same-origin as the
- *     inbound request and live under `/p/<id>/`. The cookie ships on
- *     every same-origin request so referer-based asset routing in the
- *     proxy preamble can authenticate sub-resource loads.
- *
- * This route is auth-exempt (`isAuthExempt` whitelists `/preview/login`)
- * because the standard bearer gate would 401 the browser before the
- * handler ever ran. The handler does its own bearer check via
- * `previewAuth.verifyLoginToken` (constant-time compare against the
- * configured `WARREN_API_TOKEN`).
- *
- * `redirect` is constrained to the run's own preview surface — anything
- * else is rejected so a stolen login link can't become an open redirect.
- *
- * 400 when `previewAuth` is null (subdomain mode with no host, or
- * warren booted with `--no-auth`); the proxy is also disabled in those
- * configurations so the handshake has nothing to issue against.
- */
-function previewLoginHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const runId = requireParam(ctx, "id");
-		const mode: "subdomain" | "path" = deps.previewMode ?? "subdomain";
-		if (deps.previewAuth === undefined) {
-			throw new ValidationError("preview surface is not configured on this warren", {
-				recoveryHint:
-					"ensure WARREN_API_TOKEN is set (and WARREN_PREVIEW_HOST when WARREN_PREVIEW_MODE=subdomain) to enable per-run previews",
-			});
-		}
-		if (mode === "subdomain" && deps.previewHost === undefined) {
-			throw new ValidationError("preview surface is not configured on this warren", {
-				recoveryHint:
-					"set WARREN_PREVIEW_HOST to enable subdomain-mode previews, or switch to WARREN_PREVIEW_MODE=path",
-			});
-		}
-		const token = ctx.url.searchParams.get("token");
-		if (!deps.previewAuth.verifyLoginToken(token)) {
-			return jsonResponse(401, {
-				error: {
-					code: "unauthorized",
-					message: "preview login requires a valid ?token=<WARREN_API_TOKEN>",
-				},
-			});
-		}
-		// 404 fast if the run isn't known — issuing a cookie for a nonexistent
-		// run would let an attacker pre-seed a session keyed off a future id.
-		await deps.repos.runs.require(runId);
-
-		const redirect = ctx.url.searchParams.get("redirect");
-		const redirectTarget =
-			mode === "path"
-				? resolvePathPreviewRedirect(redirect, runId, ctx.url.origin)
-				: resolveSubdomainPreviewRedirect(redirect, runId, deps.previewHost as string);
-		if (redirectTarget === null) {
-			const hint =
-				mode === "path"
-					? `redirect must be a same-origin URL under ${ctx.url.origin}/p/${runId}/`
-					: `redirect must be an absolute URL under https://run-${runId}.${deps.previewHost}/`;
-			return jsonResponse(400, {
-				error: {
-					code: "preview_redirect_invalid",
-					message: hint,
-				},
-			});
-		}
-
-		const now = deps.now?.() ?? new Date();
-		const cookie = deps.previewAuth.signCookie(runId, now);
-		return new Response(null, {
-			status: 302,
-			headers: {
-				location: redirectTarget,
-				"set-cookie": cookie.setCookieHeader,
-			},
-		});
-	};
-}
-
-function resolveSubdomainPreviewRedirect(
-	raw: string | null,
-	runId: string,
-	host: string,
-): string | null {
-	const fallback = `https://run-${runId}.${host}/`;
-	if (raw === null || raw.length === 0) return fallback;
-	let parsed: URL;
-	try {
-		parsed = new URL(raw);
-	} catch {
-		return null;
-	}
-	if (parsed.protocol !== "https:") return null;
-	if (parsed.hostname !== `run-${runId}.${host}`) return null;
-	return parsed.toString();
-}
-
-function resolvePathPreviewRedirect(
-	raw: string | null,
-	runId: string,
-	inboundOrigin: string,
-): string | null {
-	const fallback = `${inboundOrigin}/p/${runId}/`;
-	if (raw === null || raw.length === 0) return fallback;
-	let parsed: URL;
-	try {
-		// Relative URLs (`/p/<id>/foo`) resolve against the inbound origin so
-		// callers don't have to know the scheme/host upfront. Absolute URLs are
-		// then origin-checked below.
-		parsed = new URL(raw, inboundOrigin);
-	} catch {
-		return null;
-	}
-	if (parsed.origin !== inboundOrigin) return null;
-	if (!parsed.pathname.startsWith(`/p/${runId}/`)) return null;
-	return parsed.toString();
-}
-
-/**
- * `POST /runs/:id/preview/teardown` (R-19 / SPEC §11.L acceptance #8,
- * warren-d725).
- *
- * Idempotent operator-driven teardown of the per-run preview. Bearer-
- * required (the global auth gate covers `/runs/*`; this route is not
- * in `isAuthExempt`). The body is optional — `{actor}` is forwarded
- * onto the audit event for attribution, defaulting to `"manual"`.
- *
- * Responds 200 on every CAS outcome (`torn-down`, `already-torn-down`,
- * `already-failed`, `never-launched`); 404 on unknown runId; 503 when
- * `deps.db` is undefined (no repo layer wired). Works on both sqlite
- * and postgres dialects — `createRunPreviewsRepo` is dialect-
- * polymorphic (warren-adfb), so the eviction-worker CAS path that
- * teardown rides on is already exercised on pg in production. The
- * route is `tornDown: true` only when the call actually flipped a
- * `starting`/`live` row.
- */
-function previewTeardownHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const runId = requireParam(ctx, "id");
-		const body = await readJsonBodyOrEmpty(ctx);
-		const actor = body !== null ? optionalString(body, "actor") : undefined;
-
-		if (deps.db === undefined) {
-			return jsonResponse(503, {
-				error: {
-					code: "preview_teardown_unavailable",
-					message: "preview teardown requires the repo layer; this warren has no db handle wired",
-				},
-			});
-		}
-
-		const previews = createRunPreviewsRepo(deps.db);
-		const result = await teardownPreview({
-			runId,
-			repos: deps.repos,
-			previews,
-			burrowClientPool: deps.burrowClientPool,
-			broker: deps.broker,
-			...(actor !== undefined ? { actor } : {}),
-			...(deps.now !== undefined ? { now: deps.now } : {}),
-			logger: teardownLoggerFor(deps),
-		});
-
-		return jsonResponse(200, {
-			status: result.status,
-			tornDown: result.tornDown,
-			previousState: result.previousState,
-			port: result.port,
-		});
-	};
-}
-
-/**
- * Narrow `ServerDeps.logger` (the pino-shaped surface) onto the
- * `Record<string, unknown>` signature the preview teardown / eviction
- * code expects. Same shape the boot path already builds for the
- * eviction worker — kept inline here to avoid threading another
- * `*LoggerFromPino` adapter down through `ServerDeps`.
- */
-function teardownLoggerFor(deps: ServerDeps): {
-	info(obj: Record<string, unknown>, msg?: string): void;
-	warn(obj: Record<string, unknown>, msg?: string): void;
-	error(obj: Record<string, unknown>, msg?: string): void;
-} {
-	return {
-		info: (obj, msg) => deps.logger.info(obj, msg),
-		warn: (obj, msg) => deps.logger.warn(obj, msg),
-		error: (obj, msg) => deps.logger.error(obj, msg),
-	};
-}
-
-function streamRunEventsHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const id = requireParam(ctx, "id");
-		// 404 fast if the run isn't known — without this we'd happily
-		// stream an empty NDJSON forever for a typo'd id.
-		const run = await deps.repos.runs.require(id);
-
-		const follow = parseBoolean(ctx.url.searchParams.get("follow"), "follow") ?? false;
-		const sinceSeq = parseNonNegativeInt(ctx.url.searchParams.get("since"), "since");
-
-		const ctrl = bridgeAbort(ctx.request.signal);
-		const source = tailRunEvents({
-			runId: id,
-			repos: { events: deps.repos.events },
-			broker: deps.broker,
-			follow,
-			...(sinceSeq !== undefined ? { sinceSeq } : {}),
-			signal: ctrl.signal,
-		});
-		// warren-a8c3: tag every NDJSON envelope with the run's plot_id so
-		// Plot-aware consumers can route mirrored events (warren-7e0f) without
-		// a second GET /runs/:id call. Snapshot at stream-open time — plot_id
-		// is set at spawn and never mutates, so the closure-captured value is
-		// authoritative for the life of the stream.
-		const plotId = run.plotId;
-		return ndjsonResponse(asNdjsonStream(source, (row) => eventToNdjson(row, plotId), ctrl));
-	};
-}
-
-/* ----------------------------------------------------------------------- */
-/* Streaming plumbing                                                       */
-/* ----------------------------------------------------------------------- */
-
-function bridgeAbort(reqSignal: AbortSignal): AbortController {
-	const ctrl = new AbortController();
-	if (reqSignal.aborted) {
-		ctrl.abort();
-		return ctrl;
-	}
-	reqSignal.addEventListener("abort", () => ctrl.abort(), { once: true });
-	return ctrl;
-}
-
-function asNdjsonStream<T>(
-	source: AsyncIterable<T>,
-	encode: (value: T) => string,
-	ctrl: AbortController,
-): ReadableStream<Uint8Array> {
-	const encoder = new TextEncoder();
-	const iterator = source[Symbol.asyncIterator]();
-	return new ReadableStream<Uint8Array>({
-		async pull(controller) {
-			try {
-				const { done, value } = await iterator.next();
-				if (done) {
-					controller.close();
-					return;
-				}
-				controller.enqueue(encoder.encode(encode(value)));
-			} catch (err) {
-				if (ctrl.signal.aborted) {
-					controller.close();
-					return;
-				}
-				controller.error(err);
-			}
-		},
-		async cancel() {
-			ctrl.abort();
-			try {
-				await iterator.return?.(undefined);
-			} catch {
-				// ignore — generator's finally is the source of truth
-			}
-		},
-	});
-}
-
-function eventToNdjson(
-	row: {
-		id: number;
-		runId: string;
-		burrowEventSeq: number;
-		ts: string;
-		kind: string;
-		stream: string | null;
-		payloadJson: unknown;
-	},
-	plotId: string | null = null,
-): string {
-	return `${JSON.stringify({
-		id: row.id,
-		runId: row.runId,
-		seq: row.burrowEventSeq,
-		ts: row.ts,
-		kind: row.kind,
-		stream: row.stream,
-		payload: row.payloadJson,
-		plotId,
-	})}\n`;
-}
-
-/* ----------------------------------------------------------------------- */
-/* Plan runs (warren-f923 / pl-a258 step 6)                                 */
-/* ----------------------------------------------------------------------- */
-
-const PLAN_RUN_ACCEPTED_PLAN_STATUSES = ["approved", "active", "done"] as const;
-const PLAN_RUN_STATE_FILTER_VALUES = [
-	"queued",
-	"running",
-	"succeeded",
-	"failed",
-	"cancelled",
-] as const;
-type PlanRunStateFilter = (typeof PLAN_RUN_STATE_FILTER_VALUES)[number];
-
-function parsePlanRunStateFilter(raw: string | null): PlanRunStateFilter | undefined {
-	if (raw === null) return undefined;
-	if (!(PLAN_RUN_STATE_FILTER_VALUES as readonly string[]).includes(raw)) {
-		throw new ValidationError(
-			`?state must be one of ${PLAN_RUN_STATE_FILTER_VALUES.join(", ")}; got '${raw}'`,
-		);
-	}
-	return raw as PlanRunStateFilter;
-}
-
-/**
- * `POST /plan-runs` — kick off a serial plan execution against a seeds plan.
- *
- * Handler order (warren-f923):
- *   (1) load project; 404 if missing.
- *   (2) reject when project.hasSeeds is false (ProjectLacksSeedsError mirrors
- *       the plot reject shape at warren-a8c3 — same ValidationError → 400
- *       envelope, stable code so HTTP consumers can branch on it).
- *   (2b) reject when plot_id is set but project.hasPlot is false
- *       (ProjectLacksPlotError, warren-c900 / pl-7937 Phase 2). Same 400
- *       envelope; raised before the seeds-CLI fan-out so a non-Plot project
- *       never grows a half-validated plan-run.
- *
- *   Gates (2) and (2b) are **stacked, not independent** (warren-909c /
- *   pl-7937 step 6): seeds is the base, plot is the optional layer on top.
- *   A project missing .seeds/ is rejected as `project_lacks_seeds` even
- *   when plot_id is supplied — plot_id never short-circuits the .seeds/
- *   requirement, and PlanRun-with-plot only lights up when BOTH .seeds/
- *   AND .plot/ are present.
- *   (3) call showPlan; assert plan.status is in (approved, active, done) and
- *       at least one open child exists (PlanHasNoOpenChildrenError).
- *   (4) resolve agent via repos.agents.resolve with the project-tier fallback
- *       (mx-644fb5 — same posture as spawnRun).
- *   (5/6) build + persist plan_runs + plan_run_children rows in a single
- *       repo.create call (the repo runs them in a transaction so a half-
- *       inserted PlanRun never appears to listActive). plot_id rides through
- *       the same call (default null when omitted).
- *   (7) return 201 with {planRun, children}.
- */
-function createPlanRunHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const body = await readJsonBody(ctx);
-		const projectId = requireString(body, "project");
-		const planId = requireString(body, "planId");
-		const agentName = requireString(body, "agent");
-		const promptTemplate = optionalString(body, "promptTemplate");
-		const ref = optionalString(body, "ref");
-		const providerOverride = optionalString(body, "providerOverride");
-		const modelOverride = optionalString(body, "modelOverride");
-		const dispatcherHandle = optionalString(body, "dispatcherHandle");
-		const plotId = optionalString(body, "plotId");
-
-		// (1) project lookup — NotFoundError → 404.
-		const project = await deps.repos.projects.require(projectId);
-
-		// (2) hasSeeds gate.
-		if (!project.hasSeeds) {
-			throw new ProjectLacksSeedsError(
-				`project ${project.id} has no .seeds/ directory; plan-runs are not accepted`,
-				{
-					recoveryHint: "add a .seeds/ directory to the project clone and refresh",
-				},
-			);
-		}
-
-		// (2b) hasPlot gate — symmetric to single-run's spawn-time check
-		// (src/runs/spawn.ts, warren-a8c3). Empty-string plot_id is treated
-		// as "not provided" to match the single-run handler's posture.
-		if (plotId !== undefined && plotId !== "" && !project.hasPlot) {
-			throw new ProjectLacksPlotError(
-				`project ${project.id} has no .plot/ directory; plot_id is not accepted`,
-				{
-					recoveryHint:
-						"either omit plot_id on POST /plan-runs, or run `plot init` in the project clone and refresh the project so warren picks up the .plot/ directory",
-				},
-			);
-		}
-
-		// (2c) warren-bae5 / pl-5310 step 2: plot_id format + existence
-		// validation, mirroring createRunHandler's check. Layered AFTER
-		// ProjectLacksPlotError so the more-specific project-shape error
-		// still wins when both apply.
-		await assertPlotIdDispatchable({ plotId, plotResolver: deps.plotResolver });
-
-		// (3) read the plan via seeds-cli.
-		if (deps.seedsCli === undefined) {
-			throw new ValidationError(
-				"seeds CLI is not configured on this warren; plan-runs require sd",
-				{ recoveryHint: "set WARREN_SD_BINARY (or install sd on PATH) and restart" },
-			);
-		}
-		const plan = await showPlan(deps.seedsCli, project.localPath, planId);
-		if (!(PLAN_RUN_ACCEPTED_PLAN_STATUSES as readonly string[]).includes(plan.status)) {
-			throw new ValidationError(
-				`plan ${planId} is in status '${plan.status}'; plan-runs require one of ${PLAN_RUN_ACCEPTED_PLAN_STATUSES.join(", ")}`,
-				{
-					recoveryHint: "approve or activate the plan in seeds, then retry POST /plan-runs",
-				},
-			);
-		}
-		if (plan.children.length === 0) {
-			throw new PlanHasNoOpenChildrenError(`plan ${planId} has no children; nothing to dispatch`, {
-				recoveryHint: "run `sd plan submit <seed-id>` to populate the plan's children",
-			});
-		}
-		// Probe every child seed's status — if all are already closed there is
-		// nothing to dispatch. Each child is read in parallel since the seeds
-		// CLI is shell-out + filesystem read, not network.
-		const seedsCli = deps.seedsCli;
-		const childStatuses = await Promise.all(
-			plan.children.map((seedId) =>
-				showSeed(seedsCli, project.localPath, seedId).then((s) => ({
-					seedId,
-					status: s.status,
-				})),
-			),
-		);
-		const hasOpenChild = childStatuses.some((c) => c.status !== "closed");
-		if (!hasOpenChild) {
-			throw new PlanHasNoOpenChildrenError(
-				`plan ${planId} has no open children; every child seed is closed`,
-				{
-					recoveryHint: "re-open at least one child seed (sd update <id> --status open) and retry",
-				},
-			);
-		}
-
-		// (4) agent resolve with project-tier fallback (mx-644fb5).
-		const agent = await deps.repos.agents.resolve(agentName, { projectId: project.id });
-		if (agent === null) {
-			throw new NotFoundError(`agent not found: ${agentName}`, {
-				recoveryHint: "POST /agents/refresh to re-discover from canopy",
-			});
-		}
-
-		// (5/6) persist.
-		const result = await deps.repos.planRuns.create({
-			planId,
-			projectId: project.id,
-			agentName: agent.name,
-			children: plan.children.map((seedId, index) => ({ seq: index + 1, seedId })),
-			...(promptTemplate !== undefined ? { promptTemplate } : {}),
-			...(ref !== undefined ? { ref } : {}),
-			...(providerOverride !== undefined ? { providerOverride } : {}),
-			...(modelOverride !== undefined ? { modelOverride } : {}),
-			...(dispatcherHandle !== undefined ? { dispatcherHandle } : {}),
-			...(plotId !== undefined && plotId !== "" ? { plotId } : {}),
-			...(deps.now !== undefined ? { now: deps.now() } : {}),
-		});
-
-		// (6b) warren-b89f / pl-7937 step 4: emit one `plan_run_dispatched`
-		// event onto the bound Plot — fire-and-log, mirrors the single-run
-		// `defaultPlotAppender` posture in src/runs/spawn.ts:407–425. The
-		// PlanRun row is durably persisted by this point, so a Plot-write
-		// failure logs `plan_run.plot_append_failed` and the POST still
-		// returns 201.
-		if (result.planRun.plotId !== null) {
-			await emitPlanRunDispatchedToPlot({
-				appender: deps.planRunPlotAppender ?? defaultPlanRunPlotAppender,
-				logger: deps.logger,
-				plotDir: join(project.localPath, ".plot"),
-				plotId: result.planRun.plotId,
-				handle: resolveDispatcherHandle(result.planRun.dispatcherHandle),
-				planRunId: result.planRun.id,
-				planId: result.planRun.planId,
-				childrenCount: result.children.length,
-			});
-		}
-
-		// (7) wire response — coordinator picks the row up on its next tick.
-		return jsonResponse(201, {
-			planRun: result.planRun,
-			children: result.children,
-		});
-	};
-}
-
-/**
- * `GET /plan-runs?project=&state=` — list plan-runs, optionally filtered.
- * Options-bag shape mirrors `listRunsHandler` (mx-f1a881) so a UI rendering
- * either endpoint can share its query plumbing.
- */
-function listPlanRunsHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const projectId = ctx.url.searchParams.get("project");
-		const state = parsePlanRunStateFilter(ctx.url.searchParams.get("state"));
-		if (projectId !== null) {
-			const rows = await deps.repos.planRuns.listByProjectAndState(
-				projectId,
-				state !== undefined ? state : undefined,
-			);
-			return jsonResponse(200, { planRuns: rows });
-		}
-		// No project filter — return active PlanRuns when no state requested,
-		// or the operator's chosen state across every project.
-		if (state !== undefined) {
-			// listByProjectAndState requires a project; for a state-only view
-			// we walk projects sequentially. Volume is tiny relative to runs
-			// (one plan-run per dispatched plan, not per child).
-			const projects = await deps.repos.projects.listAll();
-			const all = (
-				await Promise.all(
-					projects.map((p) => deps.repos.planRuns.listByProjectAndState(p.id, state)),
-				)
-			).flat();
-			return jsonResponse(200, { planRuns: all });
-		}
-		return jsonResponse(200, { planRuns: await deps.repos.planRuns.listActive() });
-	};
-}
-
-/**
- * `GET /plan-runs/:id` — full detail page payload: row + children + the
- * fanned-out `runs[]` from runs.listByIds(child.runId for each non-null)
- * so the UI's detail page renders in one round-trip.
- */
-function getPlanRunHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const id = requireParam(ctx, "id");
-		const planRun = await deps.repos.planRuns.require(id);
-		const children = await deps.repos.planRuns.listChildren(id);
-		const runIds = children.map((c) => c.runId).filter((v): v is string => v !== null);
-		const runs = await deps.repos.runs.listByIds(runIds);
-		return jsonResponse(200, { planRun, children, runs });
-	};
-}
-
-/**
- * `POST /plan-runs/:id/cancel` — flip the plan-run to `cancelled` and, if
- * a child run is in-flight (dispatched / running / pr_open), forward a
- * cancel to that child via the existing `cancelRun` seam.
- *
- * Idempotent against an already-terminal plan-run: returns the current row
- * without firing a second cancel.
- */
-function cancelPlanRunHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const id = requireParam(ctx, "id");
-		const planRun = await deps.repos.planRuns.require(id);
-		if (
-			planRun.state === "cancelled" ||
-			planRun.state === "succeeded" ||
-			planRun.state === "failed"
-		) {
-			return jsonResponse(200, {
-				planRun,
-				cancelledChild: null,
-				alreadyTerminal: true,
-			});
-		}
-
-		const children = await deps.repos.planRuns.listChildren(id);
-		const inFlight = children.find(
-			(c) => c.state === "dispatched" || c.state === "running" || c.state === "pr_open",
-		);
-
-		const endedAt = (deps.now?.() ?? new Date()).toISOString();
-		const cancelled = await deps.repos.planRuns.transitionTo(planRun.id, "cancelled", {
-			endedAt,
-		});
-
-		let cancelledChild: { childSeq: number; runId: string } | null = null;
-		if (inFlight !== undefined && inFlight.runId !== null) {
-			try {
-				await cancelRun({
-					runId: inFlight.runId,
-					repos: deps.repos,
-					burrowClientPool: deps.burrowClientPool,
-					broker: deps.broker,
-					reason: `plan_run_cancelled:${planRun.id}`,
-					...(deps.now !== undefined ? { now: deps.now } : {}),
-					...(deps.autoOpenPr !== undefined ? { autoOpenPr: deps.autoOpenPr } : {}),
-				});
-				cancelledChild = { childSeq: inFlight.seq, runId: inFlight.runId };
-			} catch (err) {
-				// Best-effort cancel — the plan-run itself is already in the
-				// cancelled state, and the child run will land terminally via
-				// the bridge regardless. Surface the failure on the response
-				// so operators can see why the child didn't cancel cleanly.
-				deps.logger.warn(
-					{
-						planRunId: planRun.id,
-						childRunId: inFlight.runId,
-						err: err instanceof Error ? err.message : String(err),
-					},
-					"plan_run.cancel_child_failed",
-				);
-			}
-		}
-
-		return jsonResponse(200, {
-			planRun: cancelled,
-			cancelledChild,
-			alreadyTerminal: false,
-		});
-	};
-}
-
-/**
- * `GET /plan-runs/:id/events` — NDJSON tail of the union of every child
- * run's events. Read-only: snapshots `events.listByRunIds(...)` first,
- * then subscribes to the broker for each child run. Live arrivals after
- * the snapshot are deduped by (runId, burrowEventSeq).
- *
- * `?follow=1` keeps the stream open until the client disconnects or the
- * plan-run reaches a terminal state. Default (no follow) returns the
- * snapshot then closes.
- */
-function streamPlanRunEventsHandler(deps: ServerDeps): RouteHandler {
-	return async (ctx) => {
-		const id = requireParam(ctx, "id");
-		const planRun = await deps.repos.planRuns.require(id);
-		const follow = parseBoolean(ctx.url.searchParams.get("follow"), "follow") ?? false;
-		const ctrl = bridgeAbort(ctx.request.signal);
-
-		const source = tailPlanRunEvents({
-			planRun,
-			repos: deps.repos,
-			broker: deps.broker,
-			follow,
-			signal: ctrl.signal,
-		});
-		return ndjsonResponse(asNdjsonStream(source, (row) => eventToNdjson(row, null), ctrl));
-	};
-}
-
-interface TailPlanRunEventsInput {
-	readonly planRun: { id: string };
-	readonly repos: ServerDeps["repos"];
-	readonly broker: ServerDeps["broker"];
-	readonly follow: boolean;
-	readonly signal: AbortSignal;
-}
-
-type PlanRunEventRow = {
-	id: number;
-	runId: string;
-	burrowEventSeq: number;
-	ts: string;
-	kind: string;
-	stream: string | null;
-	payloadJson: unknown;
-};
-
-/**
- * Tail the union of every plan-run child's events. History first (via
- * `events.listByRunIds`), then live arrivals from `broker.subscribe(runId)`
- * for each known child runId. Newly-dispatched children are picked up by a
- * polling watcher so a stream opened before child 2 lands still sees its
- * events without a reconnect.
- *
- * Live events are deduped by (runId, burrowEventSeq) against the high-water
- * mark established during history replay, so a row that lands in the gap
- * between snapshot and subscribe isn't either dropped or duplicated.
- */
-async function* tailPlanRunEvents(
-	input: TailPlanRunEventsInput,
-): AsyncGenerator<PlanRunEventRow, void, void> {
-	const seenSeq = new Map<string, number>();
-
-	const initialChildren = await input.repos.planRuns.listChildren(input.planRun.id);
-	const initialRunIds = initialChildren.map((c) => c.runId).filter((v): v is string => v !== null);
-	const history = await input.repos.events.listByRunIds(initialRunIds);
-	for (const row of history) {
-		const prev = seenSeq.get(row.runId) ?? 0;
-		if (row.burrowEventSeq > prev) seenSeq.set(row.runId, row.burrowEventSeq);
-		yield row;
-	}
-
-	if (!input.follow) return;
-
-	// Shared event queue fed by every per-child subscription pump.
-	const queue: PlanRunEventRow[] = [];
-	let waiter: (() => void) | null = null;
-	const wake = (): void => {
-		const fn = waiter;
-		if (fn !== null) {
-			waiter = null;
-			fn();
-		}
-	};
-	input.signal.addEventListener("abort", wake, { once: true });
-
-	const subscribed = new Set<string>();
-	const subscribe = (runId: string): void => {
-		if (subscribed.has(runId)) return;
-		subscribed.add(runId);
-		const sub = input.broker.subscribe(runId, { signal: input.signal });
-		void (async () => {
-			try {
-				for await (const row of sub) {
-					queue.push(row as PlanRunEventRow);
-					wake();
-				}
-			} catch {
-				// broker.subscribe ends via signal abort or close — ignore.
-			}
-		})();
-	};
-	for (const runId of initialRunIds) subscribe(runId);
-
-	const watcherIntervalMs = 2_000;
-	const watcher = setInterval(() => {
-		void (async () => {
-			try {
-				const fresh = await input.repos.planRuns.listChildren(input.planRun.id);
-				for (const child of fresh) {
-					if (child.runId !== null) subscribe(child.runId);
-				}
-			} catch {
-				// Best-effort — a missed reload pings again next tick.
-			}
-		})();
-	}, watcherIntervalMs);
-
-	try {
-		while (!input.signal.aborted) {
-			const row = queue.shift();
-			if (row === undefined) {
-				await new Promise<void>((resolve) => {
-					waiter = resolve;
-				});
-				continue;
-			}
-			const prev = seenSeq.get(row.runId) ?? 0;
-			if (row.burrowEventSeq <= prev) continue;
-			seenSeq.set(row.runId, row.burrowEventSeq);
-			yield row;
-		}
-	} finally {
-		clearInterval(watcher);
-	}
 }
 
 /* ----------------------------------------------------------------------- */
